requirement.txt
yFlask==3.0.2
Flask-MySQLdb==2.0.0
mysql-connector-python==9.5.0
